# refactor(core): bake single-thread assumptions into spawn/spawn_blocking