--- a/cli/lsp/testing/execution.rs
+++ b/cli/lsp/testing/execution.rs
@@ -24,9 +24,11 @@ use deno_core::futures::stream;
 use deno_core::futures::StreamExt;
 use deno_core::parking_lot::Mutex;
 use deno_core::parking_lot::RwLock;
+use deno_core::task::spawn;
+use deno_core::task::spawn_blocking;
 use deno_core::ModuleSpecifier;
 use deno_runtime::permissions::Permissions;
-use deno_runtime::tokio_util::run_local;
+use deno_runtime::tokio_util::create_and_run_current_thread;
 use indexmap::IndexMap;
 use std::collections::HashMap;
 use std::collections::HashSet;
@@ -284,7 +286,7 @@ impl TestRun {
       };
       let token = self.token.clone();
 
-      tokio::task::spawn_blocking(move || {
+      spawn_blocking(move || {
         if fail_fast_tracker.should_stop() {
           return Ok(());
         }
@@ -292,13 +294,13 @@ impl TestRun {
         let file_result = if token.is_cancelled() {
           Ok(())
         } else {
-          run_local(test::test_specifier(
-            &worker_factory,
+          create_and_run_current_thread(test::test_specifier(
+            worker_factory,
             permissions,
             specifier,
             sender.clone(),
             fail_fast_tracker,
-            &test::TestSpecifierOptions {
+            test::TestSpecifierOptions {
               filter,
               shuffle: None,
               trace_ops: false,
@@ -331,7 +333,7 @@ impl TestRun {
     ));
 
     let handler = {
-      tokio::task::spawn(async move {
+      spawn(async move {
         let earlier = Instant::now();
         let mut summary = test::TestSummary::new();
         let mut used_only = false;