--- a/cli/util/draw_thread.rs
+++ b/cli/util/draw_thread.rs
@@ -2,6 +2,7 @@
 
 use console_static_text::ConsoleStaticText;
 use deno_core::parking_lot::Mutex;
+use deno_core::task::spawn_blocking;
 use deno_runtime::ops::tty::ConsoleSize;
 use once_cell::sync::Lazy;
 use std::sync::Arc;
@@ -162,7 +163,7 @@ impl DrawThread {
     internal_state.has_draw_thread = true;
 
     let drawer_id = internal_state.drawer_id;
-    tokio::task::spawn_blocking(move || {
+    spawn_blocking(move || {
       let mut previous_size = console_size();
       loop {
         let mut delay_ms = 120;