@bartlomieju 2023-05-11
Let's not land it yet, it appears this PR makes the problem in https://github.com/denoland/deno/issues/19100 even worse (ie. it takes more time to respond that that interval call). Let's investigate that issue a bit before we land, maybe it's not related, but over 1ms overhead to clamped 4ms interval seems like a serious bug.